incluindo campo telefone (fixo, celular) no cadastro de usuário, removendo tabela clients, acrescentando injeção de dependência para o cadastro de endereços do usuário, criação de actions para cadastrar, atualizar, excluir e restaurar endereços

+++ database/migrations/2016_08_04_012018_create_clients_table.php
@@ -32,6 +32,8 @@ class CreateClientsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('clients');
+        if (Schema::hasTable('clients')){
+            Schema::drop('clients');
+        }
     }
 }
